refactor(daemon): move createRemoteRuntime into k8s/remote-runtime.ts
Verbatim move of the module-level createRemoteRuntime helper out of
k8s/driver.ts into its own module, imported back by the driver. The only
import that moved with it is ShimRequestTimeoutError, used nowhere else in
driver.ts. No signature or behavior changes.
